location bearing update correction

+++ shared/src/commonMain/kotlin/com/asphalt/android/model/connectedride/ConnectedRideModel.kt
@@ -41,7 +41,7 @@ data class ConnectedRideDTO(
     val distanceTravelled: Double = 0.0,
     val rideStartedTime: Long = 0,
     var canTrack: Boolean = true,
-    var bearing: Double = 0.0
+    var bearing: Float = 0.0f
 )
 
 @Serializable
